add standalone package creation ability rearrange code into their own separate files add github actions workflow to create and auto upload the standalone tagger file update requirements.txt
