+++ README.md
@@ -6,13 +6,13 @@ Written with gpt4.
 # Setup
 
 - Clone the repo or download the zip
-- Run `setup-linux.sh` for linux or `setup-win.bat` for windows. It will insatll everything neeeded to run
+- Run `setup-linux.sh` for linux or `setup-win.bat` for windows. It will install everything needed to run
 
 # Use
 
 run `launch-linux.sh` or `launch-win.bat` to start the program.
 
-> Note: dont forget to `chmod +x` the sh files on linux
+> Note: don't forget to `chmod +x` the sh files on linux
 
 # license
